requirements.txt
Flask==3.0.2
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.1
Flask-CORS==4.0.0 